refactor: simplify index.ts to stdio-only, update test imports
Replace the monolithic index.ts with a minimal stdio-only entry point.
Update test imports to use src/server.js for createServer and
src/http.js for createApp. Adjust health endpoint assertion for the
new response shape (service + activeSessions).
